fixing routes

+++ ng-game-management/src/app/delete-game/delete-game.component.ts
@@ -19,6 +19,9 @@ export class DeleteGameComponent implements OnInit {
 
   deleteGame()
   {
+    if(confirm("Do you want to delete this game?"))
+    {
     this.service.deleteGame(this.id).subscribe((_)=>{this.router.navigate([""])});
+    }
   }
 }
